+++ app/controllers/__init__.py
@@ -1,0 +1,3 @@
+from .paper_controller import PaperController
+
+__all__ = ["PaperController"]
