+++ app/models/__init__.py
@@ -1,0 +1,3 @@
+from .paper import PaperResponse, ArxivQuery
+
+__all__ = ["PaperResponse", "ArxivQuery"]
