+++ app/services/__init__.py
@@ -1,0 +1,3 @@
+from .arxiv_service import ArxivService
+
+__all__ = ["ArxivService"]
