+++ app/utils/__init__.py
@@ -1,0 +1,3 @@
+from .config import Config
+
+__all__ = ["Config"]
